# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Aegis AI Security Platform API
  description: |
    AI/LLM 보안 플랫폼 API 명세서.

    ## 주요 기능
    - **Input Guard**: 프롬프트 인젝션 탐지
    - **Output Guard**: PII 및 민감 데이터 탐지
    - **RAG Guard**: 문서 보안 검사, 임베딩 무결성, 시맨틱 드리프트
    - **Agent Guard**: 도구 호출 검증
    - **Policy Engine**: 보안 정책 관리

    ## 인증
    현재 버전은 내부망 전용으로 인증 없이 사용 가능합니다.
    프로덕션 환경에서는 API Key 또는 JWT 인증을 권장합니다.
  version: 1.0.0
  contact:
    name: Aegis Security Team
  license:
    name: MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Local Development
  - url: https://aegis.internal/api/v1
    description: Internal Production
...
paths:
...
  /inspect:
    post:
      tags: [Input Guard]
      summary: 메시지 검사
      description: |
        입력 메시지에 대한 정밀 보안 검사 수행.
        - 패턴 기반 탐지
        - ML 기반 분류 (5-label)
        - 시맨틱 분석
        - 컨텍스트 분석
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectRequest'
      responses:
        '200':
          description: 검사 통과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectResponse'
        '403':
          description: 위협 탐지 (차단)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /output/analyze:
    post:
      tags: [Output Guard]
      summary: 출력 분석
      description: |
        LLM 출력에서 개인정보(PII) 및 민감 데이터 탐지.
        - 주민등록번호, 전화번호, 이메일 등
        - API 키, 토큰, 자격 증명
        - 내부 시스템 정보
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutputAnalyzeRequest'
      responses:
        '200':
          description: 분석 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputAnalyzeResponse'
  /rag/scan:
    post:
      tags: [RAG Guard]
      summary: 문서 스캔
      description: |
        RAG 문서 보안 검사.
        - 비가시 문자 탐지
        - 숨겨진 지시문 탐지
        - 인코딩 공격 탐지
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RAGScanRequest'
      responses:
        '200':
          description: 안전한 문서
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGScanResponse'
        '403':
          description: 위협 탐지
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGScanResponse'
  /rag/ingest:
    post:
      tags: [RAG Guard]
      summary: 배치 문서 스캔
      description: 여러 문서를 일괄 스캔
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RAGIngestRequest'
      responses:
        '200':
          description: 모든 문서 안전
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGIngestResult'
        '403':
          description: 일부 문서에서 위협 탐지
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGIngestResult'
  /rag/validate-chunks:
    post:
      tags: [RAG Guard]
      summary: 청크 검증
      description: RAG 청크 보안 검증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RAGChunkValidateRequest'
      responses:
        '200':
          description: 모든 청크 안전
        '403':
          description: 일부 청크에서 위협 탐지
  /rag/verify-embedding:
    post:
      tags: [RAG Embedding]
      summary: 임베딩 무결성 검증
      description: |
        단일 임베딩 벡터의 무결성 검증.
        - 차원 일치 확인
        - NaN/Infinity 값 탐지
        - 이상치 탐지
        - 체크섬 검증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [embedding]
              properties:
                embedding:
                  $ref: '#/components/schemas/EmbeddingVector'
                expectedDimension:
                  type: integer
                  description: 예상 차원 (선택)
      responses:
        '200':
          description: 검증 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingIntegrityResponse'
  /rag/verify-embeddings:
    post:
      tags: [RAG Embedding]
      summary: 배치 임베딩 검증
      description: 여러 임베딩 벡터 일괄 검증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [embeddings]
              properties:
                embeddings:
                  type: array
                  items:
                    $ref: '#/components/schemas/EmbeddingVector'
                expectedDimension:
                  type: integer
      responses:
        '200':
          description: 배치 검증 결과
  /rag/detect-drift:
    post:
      tags: [RAG Drift]
      summary: 시맨틱 드리프트 탐지
      description: |
        원본 콘텐츠와 현재 콘텐츠 간의 의미적 변화 탐지.
        - 토픽 변화
        - 스타일 변화
        - 인젝션 의심 패턴
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [originalContent, currentContent]
              properties:
                originalContent:
                  type: string
                currentContent:
                  type: string
      responses:
        '200':
          description: 드리프트 분석 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticDriftResponse'
  /rag/check-chunk-consistency:
    post:
      tags: [RAG Drift]
      summary: 청크 일관성 검사
      description: 문서 청크 간 시맨틱 일관성 확인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chunks]
              properties:
                chunks:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      content:
                        type: string
      responses:
        '200':
          description: 일관성 검사 결과
  /rag/content-signature:
    post:
      tags: [RAG Drift]
      summary: 콘텐츠 시그니처 생성
      description: 콘텐츠의 시맨틱 시그니처 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        '200':
          description: 콘텐츠 시그니처
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentSignatureResponse'
  /rag/provenance/create:
    post:
      tags: [RAG Provenance]
      summary: 출처 정보 생성
      description: 문서의 출처 추적 정보 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [documentId, sourceType, origin]
              properties:
                documentId:
                  type: string
                sourceType:
                  type: string
                  enum: [internal, external, user_upload, api, crawl]
                origin:
                  type: string
                content:
                  type: string
                verified:
                  type: boolean
      responses:
        '201':
          description: 출처 정보 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvenanceResponse'
  /rag/provenance/validate:
    post:
      tags: [RAG Provenance]
      summary: 출처 검증
      description: 문서 출처 정보의 무결성 검증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provenance]
              properties:
                provenance:
                  $ref: '#/components/schemas/DocumentProvenance'
                currentContent:
                  type: string
      responses:
        '200':
          description: 검증 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvenanceValidationResponse'
  /rag/provenance/check-access:
    post:
      tags: [RAG Provenance]
      summary: 접근 권한 확인
      description: 신뢰 수준 기반 접근 권한 확인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provenance, requiredTrustLevel]
              properties:
                provenance:
                  $ref: '#/components/schemas/DocumentProvenance'
                requiredTrustLevel:
                  type: string
                  enum: [verified, trusted, standard, untrusted, unknown]
      responses:
        '200':
          description: 접근 권한 결과
  /agent/validate-tool:
    post:
      tags: [Agent Guard]
      summary: 도구 호출 검증
      description: |
        에이전트의 도구 호출 요청 검증.
        - 화이트리스트 검사
        - 파라미터 검증
        - 권한 범위 확인
        - 위험도 평가
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCallValidateRequest'
      responses:
        '200':
          description: 도구 호출 허용
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCallValidateResponse'
        '403':
          description: 도구 호출 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCallValidateResponse'
  /mcp/validate:
    post:
      tags: [MCP Gateway]
      summary: MCP 요청 검증
      description: MCP 프로토콜 요청의 보안 검증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPValidateRequest'
      responses:
        '200':
          description: 요청 안전
        '403':
          description: 위협 탐지
  /llm/chat:
    post:
      tags: [LLM Proxy]
      summary: LLM 채팅 프록시
      description: |
        멀티 프로바이더 LLM 프록시.
        - OpenAI, Azure, Claude, Gemini 지원
        - 입출력 자동 검사
        - SSE 스트리밍 지원
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LLMProxyRequest'
      responses:
        '200':
          description: LLM 응답
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProxyResponse'
        '403':
          description: 요청 차단
  /policies:
    get:
      tags: [Policies]
      summary: 정책 목록 조회
      responses:
        '200':
          description: 정책 설정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyConfig'
    post:
      tags: [Policies]
      summary: 정책 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyRuleCreate'
      responses:
        '201':
          description: 정책 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyRule'
...
  /policies/rollback/{versionId}:
    post:
      tags: [Policies]
      summary: 정책 롤백
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 롤백 성공
        '404':
          description: 버전 없음
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    InspectRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: 검사할 메시지
          example: "Hello, how are you?"
        sessionId:
          type: string
          description: 세션 ID (컨텍스트 분석용)
        context:
          type: object
          description: 추가 컨텍스트 정보
    InspectResponse:
      type: object
      properties:
        requestId:
          type: string
        result:
          type: object
          properties:
            passed:
              type: boolean
            findings:
              type: array
              items:
                $ref: '#/components/schemas/Finding'
            riskScore:
              type: number
              minimum: 0
              maximum: 1
            latencyMs:
              type: integer
...